+++ src/Farmer/Builders/Builders.ContainerApps.fs
@@ -33,6 +33,7 @@ type ContainerAppConfig =
       /// Credentials for image registries used by containers in this environment.
       ImageRegistryCredentials : ImageRegistryAuthentication list
       Containers : ContainerConfig list
+      Identity : ManagedIdentity
       Dependencies : Set<ResourceId> }
 
 type ContainerEnvironmentConfig =
@@ -78,6 +79,7 @@ type ContainerEnvironmentConfig =
                   DaprConfig = containerApp.DaprConfig
                   Secrets = containerApp.Secrets
                   EnvironmentVariables = containerApp.EnvironmentVariables
+                  Identity = containerApp.Identity
                   ImageRegistryCredentials = containerApp.ImageRegistryCredentials
                   Containers = containerApp.Containers |> List.map (fun c -> c.BuildContainer)
                   Location = location
@@ -148,6 +150,7 @@ type ContainerAppBuilder () =
           Identity = ManagedIdentity.Empty
           EnvironmentVariables = Map.empty
           DaprConfig = None
+          Identity = ManagedIdentity.Empty
           Dependencies = Set.empty }
 
     member _.Run (state:ContainerAppConfig) =
@@ -164,6 +167,7 @@ type ContainerAppBuilder () =
 
         state
 
+    interface IIdentity<ContainerAppConfig> with member _.Add state updater = { state with Identity = updater state.Identity }
     /// Sets the name of the Azure Container App.
     [<CustomOperation "name">]
     member _.ResourceName (state:ContainerAppConfig, name:string) = { state with Name = ResourceName name }