+++ src/Farmer/Builders/Builders.ContainerGroups.fs
@@ -2,6 +2,7 @@
 module Farmer.Builders.ContainerGroups
 
 open Farmer
+open Farmer.Builders
 open Farmer.ContainerGroup
 open Farmer.Identity
 open Farmer.Arm.ContainerInstance
@@ -72,6 +73,10 @@ type InitContainerConfig =
 type ContainerGroupConfig =
     { /// The name of the container group.
       Name : ResourceName
+      /// Availability zone where the container group should be deployed.
+      AvailabilityZone : string option
+      /// Diagnostics and logging for the container group
+      Diagnostics : ContainerGroupDiagnostics option
       /// DNS configuration for the container group
       DnsConfig : ContainerGroupDnsConfiguration option
       /// Container group OS.
@@ -82,12 +87,16 @@ type ContainerGroupConfig =
       ImageRegistryCredentials : ImageRegistryAuthentication list
       /// IP address for the container group.
       IpAddress : ContainerGroupIpAddress option
-      /// Name of the network profile for this container's group.
-      NetworkProfile : ResourceName option
       /// The init containers in this container group.
       InitContainers : InitContainerConfig list
       /// The instances in this container group.
       Instances : ContainerInstanceConfig list
+      /// Name of the network profile for this container's group - not supported when specifying the availability zone.
+      NetworkProfile : ResourceName option
+      /// Resource ID of the virtual network where this container group should be attached.
+      VirtualNetwork : LinkedResource option
+      /// Name of the subnet where this container group should be attached.
+      SubnetName : ResourceName option
       /// Volumes to mount on the container group.
       Volumes : Map<string, Volume>
       /// Managed identity for the container group.
@@ -103,6 +112,11 @@ type ContainerGroupConfig =
         member this.BuildResources location = [
             { Location = location
               Name = this.Name
+              AvailabilityZone =
+                  if this.AvailabilityZone.IsSome && this.NetworkProfile.IsSome then
+                      raiseFarmer $"Cannot specify availability zone when using network profiles."
+                  else
+                    this.AvailabilityZone
               ContainerInstances = [
                 for instance in this.Instances do
                     match instance.Image with
@@ -120,6 +134,7 @@ type ContainerGroupConfig =
                            ReadinessProbe = instance.ReadinessProbe
                            VolumeMounts = instance.VolumeMounts |}
               ]
+              Diagnostics = this.Diagnostics
               DnsConfig =
                   if this.DnsConfig.IsSome && this.NetworkProfile.IsNone then
                     raiseFarmer "DNS configuration can only be set when attached to a virtual network."
@@ -140,7 +155,19 @@ type ContainerGroupConfig =
                            VolumeMounts = initContainer.VolumeMounts |}
               ]
               IpAddress = this.IpAddress
-              NetworkProfile = this.NetworkProfile
+              NetworkProfile =
+                  if this.NetworkProfile.IsSome && (this.VirtualNetwork.IsSome || this.SubnetName.IsSome) then
+                      raiseFarmer $"Should not set network profile on container group '{this.Name.Value}' when using vnet and subnet."
+                  else this.NetworkProfile
+              SubnetIds =
+                  match this.VirtualNetwork, this.SubnetName with
+                  | None, None -> []
+                  | Some (Managed vnetId), Some subnet ->
+                      { vnetId with Type = subnets; Segments = [ subnet ] } |> Managed |> List.singleton
+                  | Some (Unmanaged vnetId), Some subnet ->
+                      { vnetId with Type = subnets; Segments = [ subnet ] } |> Unmanaged |> List.singleton
+                  | Some vnetId, None -> raiseFarmer $"Missing subnet for attaching container group '{this.Name.Value}' to vnet '{vnetId.Name.Value}'."
+                  | None, subnetName -> raiseFarmer $"Missing vnet for attaching container group '{this.Name.Value}' to subnet '{subnetName.Value}'."
               Volumes = this.Volumes
               Tags = this.Tags
               Dependencies = this.Dependencies }
@@ -196,6 +223,7 @@ type ContainerGroupBuilder() =
 
     member _.Yield _ =
         { Name = ResourceName.Empty
+          Diagnostics = None
           DnsConfig = None
           OperatingSystem = Linux
           RestartPolicy = AlwaysRestart
@@ -204,8 +232,11 @@ type ContainerGroupBuilder() =
           InitContainers = []
           IpAddress = None
           NetworkProfile = None
+          SubnetName = None
+          VirtualNetwork = None
           Instances = []
           Volumes = Map.empty
+          AvailabilityZone = None
           Tags = Map.empty
           Dependencies = Set.empty }
     member this.Run (state:ContainerGroupConfig) =
@@ -216,8 +247,8 @@ type ContainerGroupBuilder() =
 
     member this.AddTcpPort(state:ContainerGroupConfig, port) = this.AddPort (state, TCP, port)
 
-    [<CustomOperation "name">]
     /// Sets the name of the container group.
+    [<CustomOperation "name">]
     member _.Name(state:ContainerGroupConfig, name) = { state with Name = name }
     member this.Name(state:ContainerGroupConfig, name) = this.Name(state, ResourceName name)
     /// Sets the OS type (default Linux)
@@ -242,6 +273,19 @@ type ContainerGroupBuilder() =
     [<CustomOperation "network_profile">]
     member _.NetworkProfile(state:ContainerGroupConfig, networkProfileName:string) = { state with NetworkProfile = Some (ResourceName networkProfileName) }
     member _.NetworkProfile(state:ContainerGroupConfig, networkProfile:NetworkProfileConfig) = { state with NetworkProfile = Some networkProfile.Name }
+    /// Sets the name of a virtual network where this container group should be attached.
+    [<CustomOperation "vnet">]
+    member _.VNetId(state:ContainerGroupConfig, vnetId:ResourceId) = { state with VirtualNetwork = Some (Managed vnetId) }
+    member _.VNetId(state:ContainerGroupConfig, vnetName:string) = { state with VirtualNetwork = Some (Managed (virtualNetworks.resourceId (ResourceName vnetName))) }
+    member _.VNetId(state:ContainerGroupConfig, vnetName:ResourceName) = { state with VirtualNetwork = Some (Managed (virtualNetworks.resourceId vnetName)) }
+    /// Sets the name of a virtual network where this container group should be attached.
+    [<CustomOperation "link_to_vnet">]
+    member _.LinkToVNetId(state:ContainerGroupConfig, vnetId:ResourceId) = { state with VirtualNetwork = Some (Unmanaged vnetId) }
+    member _.LinkToVNetId(state:ContainerGroupConfig, vnetName:string) = { state with VirtualNetwork = Some (Unmanaged (virtualNetworks.resourceId (ResourceName vnetName))) }
+    member _.LinkToVNetId(state:ContainerGroupConfig, vnetName:ResourceName) = { state with VirtualNetwork = Some (Unmanaged (virtualNetworks.resourceId vnetName)) }
+    [<CustomOperation "subnet">]
+    member _.Subnet(state:ContainerGroupConfig, subnetName:string) = { state with SubnetName = Some (ResourceName subnetName) }
+    member _.Subnet(state:ContainerGroupConfig, subnetName:ResourceName) = { state with SubnetName = Some subnetName }
     /// Adds a UDP port to be externally accessible
     [<CustomOperation "add_udp_port">]
     member this.AddUdpPort(state:ContainerGroupConfig, port) = this.AddPort (state, UDP, port)
@@ -259,12 +303,51 @@ type ContainerGroupBuilder() =
     /// Adds a collection of container instances to this group
     [<CustomOperation "add_instances">]
     member _.AddInstances(state:ContainerGroupConfig, instances) = { state with Instances = state.Instances @ (Seq.toList instances) }
-    [<CustomOperation "add_volumes">]
     /// Adds volumes to the container group so they can be mounted on containers.
+    [<CustomOperation "add_volumes">]
     member _.AddVolumes(state:ContainerGroupConfig, volumes) =
         let newVolumes = volumes |> Map.ofSeq
         let updatedVolumes = state.Volumes |> Map.fold (fun current key vol -> Map.add key vol current) newVolumes
         { state with Volumes = updatedVolumes }
+    /// Specify the availability zone for the container group.
+    [<CustomOperation "availability_zone">]
+    member _.AvailabilityZones(state:ContainerGroupConfig, zone:string) =
+        { state with AvailabilityZone = Some zone }
+    [<CustomOperation "diagnostics_workspace">]
+    member _.EnableDiagnostics(state:ContainerGroupConfig, logType:LogType, workspaceBuilder:WorkspaceConfig) =
+        { state with
+            Diagnostics =
+                {
+                    LogType = logType
+                    Workspace = LogAnalyticsWorkspace.WorkspaceResourceId(Managed((workspaceBuilder :> IBuilder).ResourceId))
+                } |> Some
+        }
+    member _.EnableDiagnostics(state:ContainerGroupConfig, logType:LogType, workspaceResourceId:ResourceId) =
+        { state with
+            Diagnostics =
+                {
+                    LogType = logType
+                    Workspace = LogAnalyticsWorkspace.WorkspaceResourceId(Managed(workspaceResourceId))
+                } |> Some
+        }
+    [<CustomOperation "diagnostics_workspace_key">]
+    member _.EnableDiagnosticsWorkspace(state:ContainerGroupConfig, logType:LogType, workspaceId:string, workspaceKey:string) =
+        { state with
+            Diagnostics =
+                {
+                    LogType = logType
+                    Workspace = LogAnalyticsWorkspace.WorkspaceKey(workspaceId, workspaceKey)
+                } |> Some
+        }
+    [<CustomOperation "link_to_diagnostics_workspace">]
+    member _.LinkToDiagnosticsWorkspace(state:ContainerGroupConfig, logType:LogType, workspaceResourceId:ResourceId) =
+        { state with
+            Diagnostics =
+                {
+                    LogType = logType
+                    Workspace = LogAnalyticsWorkspace.WorkspaceResourceId(Unmanaged(workspaceResourceId))
+                } |> Some
+        }
     /// Specify DNS nameservers for the containers in the container group.
     [<CustomOperation "dns_nameservers">]
     member _.DnsNameServers(state:ContainerGroupConfig, nameServers:string list) =